Add files via upload
Data sets show how Ohio counties voted in 2016 and 2020. Another dataset shows how young minorities view the presidential candidates for the 2020 election. I want this data to tell or show me how counties voted differently in 2016 vs 2020. I also want the data to show me how minority groups view the candidates. In order to tidy the dataset, I had to delete blank cells and rows, add a value of 0 to blank cells so that it doesn't appear as NaN on jupyter notebook, and make sure the rows are even and don't have spaces in between them. From these data sets, I found that some Ohio counties flipped in this election and changed how they voted from the 2016 election. I also found out that young Black Americans don't approve of Biden or Trump.  This data supports my final project because it shows the differences in how counties and people voted in Ohio from the 2016 election to the 2020 election.